+++ packages/gatsby-source-wordpress-experimental/src/models/gatsby-api.js
@@ -1,6 +1,7 @@
 import merge from "lodash/merge"
 import { createRemoteMediaItemNode } from "~/steps/source-nodes/create-nodes/create-remote-media-item-node"
 import { menuBeforeChangeNode } from "~/steps/source-nodes/before-change-node/menu"
+import { categoryBeforeChangeNode } from "~/steps/source-nodes/before-change-node/category"
 
 const defaultPluginOptions = {
   url: null,
@@ -84,6 +85,10 @@ const defaultPluginOptions = {
     ContentNode: {
       nodeInterface: true,
     },
+    Category: {
+      // @todo remove this when categories are a flat list in WPGQL
+      beforeChangeNode: categoryBeforeChangeNode,
+    },
     Menu: {
       /**
        * This is used to fetch child menu items
@@ -99,6 +104,7 @@ const defaultPluginOptions = {
        *
        * When we can get a list of all menu items regardless of location in WPGQL, this can be removed.
        */
+      // @todo remove this when menus are a flat list in WPGQL
       beforeChangeNode: menuBeforeChangeNode,
     },
     MenuItem: {