+++ packages/gatsby-source-wordpress-experimental/src/supported-remote-plugin-versions.js
@@ -1,0 +1,13 @@
+const supportedWpPluginVersions = {
+  WPGraphQL: {
+    version: `~0.8.2`,
+  },
+  WPGatsby: {
+    version: `~0.2.3`,
+  },
+}
+
+// @todo replace this link with another once we're out of alpha
+const genericDownloadMessage = `\n\n\tVisit https://github.com/TylerBarnes/using-gatsby-source-wordpress-experimental/tree/master/WordPress/plugins/\n\tto download the latest supported versions of WPGatsby and WPGraphL.`
+
+export { supportedWpPluginVersions, genericDownloadMessage }