+++ packages/wp-gatsby/wp-gatsby.php
@@ -2,7 +2,7 @@
 /**
  * Plugin Name: WP Gatsby
  * Description: Optimize your WordPress site to be a source for Gatsby site(s).
- * Version: 0.2.2
+ * Version: 0.2.3
  * Author: GatsbyJS, Jason Bahl, Tyler Barnes
  * Author URI: https://gatsbyjs.org
  * Text Domain: wp-gatsby
@@ -91,7 +91,7 @@ final class WPGatsby {
 	private function setup_constants() {
 		// Plugin version.
 		if ( ! defined( 'WPGATSBY_VERSION' ) ) {
-			define( 'WPGATSBY_VERSION', '0.2.2' );
+			define( 'WPGATSBY_VERSION', '0.2.3' );
 		}
 
 		// Plugin Folder Path.